Add in legitimate rspec test for one specific sub-scenario
