+++ lib/ca_testing/extensions.rb
@@ -2,6 +2,6 @@
 
 require "ca_testing/extensions/capybara/node/element"
 require "ca_testing/extensions/cucumber/core/test/case"
-require "ca_testing/extensions/selenium/webdriver/driver_extensions/has_logs_two"
+require "ca_testing/extensions/selenium/webdriver/logs"
 require "ca_testing/extensions/array"
 require "ca_testing/extensions/string"
